requirements.txt
python-dotenv==1.0.0
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1
google.generativeai==0.3.1
streamlit==1.29.0
networkx==3.2.1
pyvis==0.3.2
pandas==2.1.4 